requirements.txt
absl-py==1.2.0
alabaster==0.7.12
alembic==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.12.11
attrs==22.1.0
autopage==0.5.1
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
cachetools==5.2.0
certifi
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cliff==3.10.1
cloudpickle==2.2.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.5
colorlog==6.7.0
commonmark==0.9.1
contextlib2==21.6.0
coverage==6.5.0
cucim==22.8.1
cycler==0.11.0
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dicom2nifti==2.4.3
dill==0.3.5.1
distlib==0.3.6
docker==6.0.0
docutils==0.16
dominate==2.7.0
einops==0.5.0
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.8.0
fire==0.4.0
flake8==5.0.4
flake8-bugbear==22.9.23
flake8-comprehensions==3.10.0
flake8-executable==2.1.2
Flask==2.2.2
fonttools==4.37.4
future==0.18.2
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.12.0
google-auth-oauthlib==0.4.6
greenlet==1.1.3.post0
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
huggingface-hub==0.10.1
hyperopt==0.1.2
identify==2.5.6
idna==3.3
imagecodecs==2021.11.20
imageio==2.21.2
imagesize==1.4.1
importlab==0.8
importlib-metadata==4.2.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==7.34.0
ipython-genutils==0.2.0
isort==5.10.1
itk==5.2.1.post1
itk-core==5.2.1.post1
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json-tricks==3.15.5
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
libcst==0.4.7
lmdb==1.3.0
Mako==1.2.3
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mlflow==1.29.0
monai==0.9.1
mypy==0.982
mypy-extensions==0.4.3
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==4.0.1
ninja==1.10.2.4
nni==2.9
nodeenv==1.7.0
notebook==6.4.12
nptyping==2.3.1
numpy==1.21.5
oauthlib==3.2.1
openslide-python==1.1.2
optuna==3.0.3
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pathspec==0.10.1
pbr==5.10.0
pep8-naming==0.13.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pprint36==3.9.0.2
pre-commit==2.20.0
prettytable==3.4.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydicom==2.3.0
pydot==1.4.2
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.5.0
pylint==2.15.4
pymongo==4.2.0
pynrrd==1.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-gdcm==3.0.17
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.10
pytype==2022.9.27
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
querystring-parser==1.2.4
recommonmark==0.6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.22.0
rsa==4.9
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
SimpleITK==2.2.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==3.5.3
sphinx-autodoc-typehints==1.11.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.41
sqlparse==0.4.3
stevedore==3.5.0
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
transformers==4.21.3
typed-ast==1.5.4
typeguard==2.13.3
types-pkg-resources==0.1.3
types-PyYAML==6.0.12
types-toml==0.10.8
typing-inspect==0.8.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.2
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0